requirements.txt
gdown==4.2.1
got10k==0.1.3
h5py==3.1.0
imgaug==0.4.0
kornia==0.5.11
numpy==1.19.2
opencv-python==4.5.5.62
pycocotools==2.0.4
pytorch-lightning==1.5.6
tqdm==4.62.3

# === Required by armory / ART === #
apache-beam==2.22.0
boto3==1.17.20
docker==5.0.3
ffmpeg-python==0.2.0
pandas==1.2.4
Pillow==8.3.0
pytest==7.0.1
requests==2.27.1
tensorflow==2.5.1
tensorflow-datasets==3.2.0

# === Required by PySOT === #
cython==0.29.27
pyyaml==6.0
tensorboardx==2.4.1
yacs==0.1.8

# === Development dependencies === #
# ipykernel==6.9.0
# mypy==0.921
# torch-tb-profiler==0.3.1
